templates added + post&get requests ready + README

+++ src/main/java/com/example/springkpi/lab2/models/ExchangeRate.java
@@ -10,8 +10,26 @@ import java.time.LocalDate;
 @Data
 @RequiredArgsConstructor
 public class ExchangeRate {
-    private int id;
+    private Long id;
     private Currency currency;
     private LocalDate date;
     private BigDecimal rate;
+//    public ExchangeRate(Long id, Currency currency, BigDecimal rate, LocalDate date) {
+//        this.id = id;
+//        this.currency = currency;
+//        this.rate = rate;
+//        this.date = date;
+//    }
+//
+//    public Long getId() { return id; }
+//    public void setId(Long id) { this.id = id; }
+//
+//    public Currency getCurrency() { return currency; }
+//    public void setCurrency(Currency currency) { this.currency = currency; }
+//
+//    public BigDecimal getRate() { return rate; }
+//    public void setRate(BigDecimal rate) { this.rate = rate; }
+//
+//    public LocalDate getDate() { return date; }
+//    public void setDate(LocalDate date) { this.date = date; }
 }
